utils,cbor: Use uint64_t for type
new_reference has code that expects the emitted field to be sometimes
larger than a 32bit value so it's necessary to make the type 64 bits on
32-bit architectures.
Also reordered the field in the struct to likely get slightly better
packing efficiency.

+++ libutils/src/cbor64.c
@@ -158,7 +158,7 @@ static int new_reference(base64_t *streamer, cbor64_domain_t *domain, size_t ind
      * encoding, don't actually track the string in the references.
      */
     size_t length = strlen(domain->strings[index]);
-    size_t next_ref = domain->emitted;
+    uint64_t next_ref = domain->emitted;
 
     bool is_referenced = true;
     if (next_ref < 24) {
